+++ niryo_robot_hardware_stack/ttl_driver/include/ttl_driver/end_effector_driver.hpp
@@ -81,6 +81,7 @@ class EndEffectorDriver : public AbstractEndEffectorDriver
         int readCollisionStatus(uint8_t id, bool& status) override;
 
         int readDigitalInput(uint8_t id, bool& in) override;
+        int writeDigitalInput(uint8_t id, bool in) override;
         int writeDigitalOutput(uint8_t id, bool out) override;
 
         int writeCollisionThresh(uint8_t id, int thresh) override;
@@ -438,6 +439,18 @@ int EndEffectorDriver<reg_type>::readDigitalInput(uint8_t id, bool& in)
     int res = read<typename reg_type::TYPE_DIGITAL_IN>(reg_type::ADDR_DIGITAL_IN, id, value);
     in = (value > 0) ? true : false;
     return res;
+}
+
+/**
+ * @brief EndEffectorDriver<reg_type>::writeDigitalInput not implemented for real end effectors
+ * @param id the device ID
+ * @param out the output value to write
+ * @return the result of the operation as a status code
+ */
+template<typename reg_type>
+int EndEffectorDriver<reg_type>::writeDigitalInput(uint8_t id, bool in)
+{
+    return COMM_TX_ERROR;
 }
 
 /**